Working sequelize DB, refactored to be more modular
