+++ server/database/resetDB.sql
@@ -1,0 +1,6 @@
+DROP DATABASE if EXISTS options;
+create DATABASE options;
+USE options;
+
+-- to run script run:
+-- mysql -u root < database/resetDB.sql
